complaint_agent/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
pydantic==2.7.1
nltk==3.8.1
spacy==3.7.4
langchain==0.2.1
langchain-openai==0.1.8
openai==1.30.1
python-dotenv==1.0.1
